% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemdatafunctions.R
\name{agrWeek}
\alias{agrWeek}
\title{Aggregate chemistry data variable by week of the year in order to compare across sites}
\usage{
agrWeek(x)
}
\arguments{
\item{x}{An object from getChem()}
}
\value{
A list of 2 column dataframes for each year at a core site containing week of year vs average weekly concentrations
}
\description{
Aggregate chemistry data variable by week of the year in order to compare across sites
}
\examples{
agrWeek(getChem(s = "GBBB", Year = "2017", Site = 16, Var = "Wtemp_C"))
}
\keyword{chemistry}
\keyword{data}
