% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpfunctions.R
\name{avgBVP}
\alias{avgBVP}
\title{Find average sonde data from a given range of depths off the lake bottom}
\usage{
avgBVP(L, D)
}
\arguments{
\item{L}{A list created from the function vpList}

\item{D}{Depths from the lake bottom to average}
}
\value{
Returns a list of vectors with average sonde data variable over the depth range from the bottom up to D meters from the bottom
}
\description{
Find average sonde data from a given range of depths off the lake bottom
}
\examples{
avgBVP(L = vpList(Var = vpVars()[1]), D = 2)
}
\keyword{data}
\keyword{sonde}
