% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpfunctions.R
\name{avgVP}
\alias{avgVP}
\title{Find average sonde data over all depths}
\usage{
avgVP(L)
}
\arguments{
\item{L}{A list created from the function vpList}
}
\value{
Returns a list of vectors with average sonde data of the chosen variable over all depths
}
\description{
Find average sonde data over all depths
}
\examples{
avgVP(L = vpList(Var = vpVars()[1], vpSites()[1:3]))
}
\keyword{data}
\keyword{sonde}
