% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blooms.R
\name{bloomDetect}
\alias{bloomDetect}
\title{Detect blooms based on a variable (expects Phy_Spec) two standard deviations above the mean and at least 50 or higher}
\usage{
bloomDetect(x)
}
\arguments{
\item{x}{A list of dataframes inherited from agrWeek(getChem())}
}
\value{
A list of 2- column dataframes for each year at one site containing sample dates and var concentrations
}
\description{
Detect blooms based on a variable (expects Phy_Spec) two standard deviations above the mean and at least 50 or higher
}
\examples{
bloomDetect(x)
}
\keyword{chemistry}
\keyword{data}
