% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{dataframe}
\name{Chem}
\alias{Chem}
\title{Water chemistry data from all sites in the Green Bay HABs study 2016 - 2020}
\format{
A data frame object.
}
\source{

}
\usage{
Chem
}
\description{
Water chemistry data from 2016 through 2020 at ALL sampling sites in Green Bay
}
\examples{
LazyData: true

}
\references{

}
\keyword{datasets}
