% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ChemPrime}
\alias{ChemPrime}
\title{Water chemistry data from common core sites in the Green Bay HABs study 2016 - 2020}
\format{
A data frame object.
}
\source{

}
\usage{
ChemPrime
}
\description{
Water chemistry data 2016 through 2020 at the six core sampling sites in Green Bay
}
\examples{
LazyData: true

}
\references{

}
\keyword{datasets}
