% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ChemVars}
\alias{ChemVars}
\title{Water Chemistry Variables}
\format{
A character vector
}
\source{

}
\usage{
ChemVars
}
\description{
Water chemistry variables measured in the Green Bay HABs study, 2016 - 2020
}
\examples{

}
\references{

}
\keyword{datasets}
