% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemdatafunctions.R
\name{corChemAvg}
\alias{corChemAvg}
\title{Perform a correlation between the weekly average acros all sites of two ChemVars variables}
\usage{
corChemAvg(Var1, Var2, Years, Sites, method = "pearson")
}
\arguments{
\item{Var1}{First variable to include in the correlation, from ChemVars - single character name}

\item{Var2}{Second variable to include in the correlation, from ChemVars - single character name}

\item{Years}{Which years to include in averaging Var1 and Var2 - character vector}

\item{Sites}{Which sites to include in the average, from names(ChemPrime) - character vector}

\item{method}{Method to use for the correlation, from ?cor.test - single character name}
}
\value{
Results from cor.test
}
\description{
Perform a correlation between the weekly average acros all sites of two ChemVars variables
}
\examples{
getChemAvg(Sites = names(ChemPrime),Var = ChemVars[11])
}
\keyword{chemistry}
\keyword{data}
