% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{freqL}
\alias{freqL}
\title{Calculate the percentage of values above a minimum}
\usage{
freqL(x, l)
}
\arguments{
\item{x}{a vector}

\item{l}{the minimum threshold number, a single number}
}
\value{
A single number representing the frequency above l as a percentage
}
\description{
Calculate the percentage of values above a minimum
}
\examples{
freqL(x)
}
\keyword{stats}
