% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemdatafunctions.R
\name{getChem}
\alias{getChem}
\title{Get water chemistry data for a given variable from core sampling sites for plotting purposes}
\usage{
getChem(s, v)
}
\arguments{
\item{s}{A core sampling site, run names(ChemPrime)}

\item{v}{A chemistry var or variable to get, see ChemVars}
}
\value{
A list of 2- column dataframes for each year at one site containing sample dates and var concentrations
}
\description{
Get water chemistry data for a given variable from core sampling sites for plotting purposes
}
\examples{
getChem(s = "GBBB", Year = "2017", Site = 16, Var = "Wtemp_C")
}
\keyword{chemistry}
\keyword{data}
