% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chemdatafunctions.R
\name{getChemAvg}
\alias{getChemAvg}
\title{Get the average, standard deviation and error of chemical variable across sites by week in each year of the Green Bay HABs study}
\usage{
getChemAvg(Sites, Var, Years)
}
\arguments{
\item{Sites}{Which sites to include in the average, from names(ChemPrime) - character vector}

\item{Var}{Which variable to average, from ChemVars - single character name}

\item{Years}{Which years to average, from Years object - character vector}
}
\value{
A list of four column dataframes for each year with average weekly concentrations across sampling sites
}
\description{
Get the average, standard deviation and error of chemical variable across sites by week in each year of the Green Bay HABs study
}
\examples{
getChemAvg(Sites = names(ChemPrime),Var = ChemVars[11])
}
\keyword{chemistry}
\keyword{data}
