% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phytofunctions.R
\name{GetCounts}
\alias{GetCounts}
\title{Get cell counts for a given site}
\usage{
GetCounts(
  Mat,
  Site,
  Locs = "Sample.Location",
  Taxa = "DNR.Parameter.Description",
  Dates = "Sample.Collected..Start..Date",
  Units = "Units",
  UnitType = "cells/mL",
  Result = "Result.Amt"
)
}
\arguments{
\item{Mat}{Phyto data frame or a dataframe subsetted from Phyto}

\item{Site}{One of the sampling sites from unique(Phyto$Sample.Location)}

\item{Taxa}{Defaults to "DNR.Parameter.Description"}

\item{Dates}{Defaults to "Sample.Collected..Start..Date",}

\item{Units}{Defaults to "Units"}

\item{UnitType}{Default is "cells/ml" but could be either "cells/mL" or "NU/ml" where NU/ml is natural units such as filaments or colonies}

\item{Result}{Defaults to "Result.Amt"}
}
\value{
A data frame of cell densities by sample date and species
}
\description{
Get cell counts for a given site
}
\examples{
GetCounts(Mat = Phyto, Site = unique(Phyto$Sample.Location)[1])
}
\keyword{phytoplankton}
