% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{getGEO}
\alias{getGEO}
\title{Calculate the geometric mean}
\usage{
getGEO(x)
}
\arguments{
\item{x}{a vector}
}
\value{
A single number representing the geometric mean of x
}
\description{
Calculate the geometric mean
}
\examples{
getGEO(x)
}
\keyword{stats}
