% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpfunctions.R
\name{getTDiff}
\alias{getTDiff}
\title{Find average temperature difference between lake surface and bottom.}
\usage{
getTDiff(Year, VP, Site)
}
\arguments{
\item{Year}{A year of the study as character one of "2016" through "2020"}

\item{VP}{Green Bay vertical profile data frame}

\item{Site}{A sampling site as given by function vpSites()}
}
\value{
Returns a vector of difference in temperature from surface minus bottom each sampling day at the site indicated. It averages top two depths of 0 and 1 m and two bottom depths then subtracts
}
\description{
Find average temperature difference between lake surface and bottom.
}
\examples{
getTDiff(VP, Year = "2017", Site = 16)
}
\keyword{data}
\keyword{sonde}
