% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpfunctions.R
\name{getVP}
\alias{getVP}
\title{Get vertical profile data for a sonde parameter}
\usage{
getVP(VP, Year, Site, Var)
}
\arguments{
\item{VP}{Green Bay vertical profile data frame}

\item{Year}{A year of the study as character one of "2016" through "2020"}

\item{Site}{A sampling site as given by function vpSites()}

\item{Var}{A sonde data variable to get as listed in the function vpVars()}
}
\value{
A data frame of variable data by depth (meters) in rows and date as columns
}
\description{
Get vertical profile data for a sonde parameter
}
\examples{
getVP(VP, Year = "2017", Site = 16, Var = Wtemp_C)
}
\keyword{data}
\keyword{sonde}
