% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{maxNA}
\alias{maxNA}
\title{Calculate the maximum after removing NA values}
\usage{
maxNA(x)
}
\arguments{
\item{x}{a vector}
}
\value{
A single number representing the maximum
}
\description{
Calculate the maximum after removing NA values
}
\examples{
maxNA(x)
}
\keyword{stats}
