% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{meanNA}
\alias{meanNA}
\title{Calculate a mean after removing NA values}
\usage{
meanNA(x)
}
\arguments{
\item{x}{a vector}
}
\value{
A single number representing the average
}
\description{
Calculate a mean after removing NA values
}
\examples{
meanNA(x)
}
\keyword{stats}
