% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{minNA}
\alias{minNA}
\title{Calculate the minimum after removing NA values}
\usage{
minNA(x)
}
\arguments{
\item{x}{a vector}
}
\value{
A single number representing the minimum
}
\description{
Calculate the minimum after removing NA values
}
\examples{
minNA(x)
}
\keyword{stats}
