% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{dataframe}
\name{Phyto}
\alias{Phyto}
\title{Phytoplankton cell count data from the Green Bay HABs study 2016 - 2018}
\format{
A data frame object.
}
\source{

}
\usage{
Phyto
}
\description{
Phytoplankton cell count data from Green Bay at select sites 2016 - 2018
}
\examples{
LazyData: true

}
\references{

}
\keyword{datasets}
