% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{sdNA}
\alias{sdNA}
\title{Calculate the standard deviation after removing NA values}
\usage{
sdNA(x)
}
\arguments{
\item{x}{a vector}
}
\value{
A single number representing the standard deviation
}
\description{
Calculate the standard deviation after removing NA values
}
\examples{
sdNA(x)
}
\keyword{stats}
