% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{SE}
\alias{SE}
\title{Calculate the standard error of the mean after removing NA values}
\usage{
SE(x)
}
\arguments{
\item{x}{a vector}
}
\value{
A single number representing the standard error
}
\description{
Calculate the standard error of the mean after removing NA values
}
\examples{
maxNA(x)
}
\keyword{stats}
