% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phytofunctions.R
\name{taxaAvg}
\alias{taxaAvg}
\title{Get the average cell density of a phylum across all sites}
\usage{
taxaAvg(phylum, sites)
}
\arguments{
\item{phylum}{A single phylum to average, given by unique(Phyto$PHYLUM.algaebase.org)}

\item{sites}{Which sites to use in the average given by unique(Phyto$Sample.Location). Multiple sites expected.}
}
\value{
A data frame of cell densities by sample date and species
}
\description{
Get the average cell density of a phylum across all sites
}
\examples{
taxaAvg(sites = unique(Phyto$Sample.Location),phylum = unique(Phyto$PHYLUM.algaebase.org)[1])
}
\keyword{phytoplankton}
