% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VP}
\alias{VP}
\title{Green Bay Vertical Profile Sonde Data}
\format{
A data frame object.
}
\source{

}
\usage{
VP
}
\description{
Sonde data from 2016 through 2020 at sampling sites in Green Bay
}
\examples{
LazyData: true

}
\references{

}
\keyword{datasets}
