% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpfunctions.R
\name{vpList}
\alias{vpList}
\title{Get all data for a sonde parameter}
\usage{
vpList(Var, VPSites)
}
\arguments{
\item{Var}{The sonde parameter of interest}

\item{VPSites}{The sampling sites of interest as given by vpSites()}
}
\value{
Returns a list of data frames, one for each year of the study with the Var chosen by depth and date
}
\description{
Get all data for a sonde parameter
}
\examples{
vpList(Var = vpVars()[1])
}
\keyword{data}
\keyword{sonde}
