% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Years}
\alias{Years}
\title{Years of the Green Bay Study}
\format{
A character vector
}
\source{

}
\usage{
Years
}
\description{
Years of the Green Bay study, 2016 - 2020
}
\examples{
LazyData: true

}
\references{

}
\keyword{datasets}
